+++ packages/expo-location/ios/Requesters/EXBackgroundLocationPermissionRequester.m
@@ -151,7 +151,15 @@ static SEL alwaysAuthorizationSelector;
     }
   }
   
-  return @{ @"status": @(status), @"scope": @(systemStatus == kCLAuthorizationStatusAuthorizedWhenInUse ? "whenInUse" : systemStatus == kCLAuthorizationStatusAuthorizedAlways ? "always" : "none") };
+  NSString *accuracy = @"full";
+  if (@available(iOS 14, *)) {
+    CLLocationManager *manager = [[CLLocationManager alloc] init];
+    if (manager.accuracyAuthorization == CLAccuracyAuthorizationReducedAccuracy) {
+      accuracy = @"reduced";
+    }
+  }
+
+  return @{ @"status": @(status), @"scope": @(systemStatus == kCLAuthorizationStatusAuthorizedWhenInUse ? "whenInUse" : systemStatus == kCLAuthorizationStatusAuthorizedAlways ? "always" : "none"), @"accuracy": accuracy };
 }
 
 @end