+++ packages/expo-location/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### 🎉 New features
 
+- [iOS] Added `accuracy` field to `PermissionDetailsLocationIOS` exposing iOS 14+ precise/reduced location authorization. ([#XXXX](https://github.com/expo/expo/pull/XXXX) by [@sbs44](https://github.com/sbs44))
+
 ### 🐛 Bug fixes
 
 ### 💡 Others